==> food_access/code/02_efap_geocode.ipynb <==
import numpy as np
import pandas as pd
import geopandas as gpd
import googlemaps

# API from NM work gmail
gmaps = googlemaps.Client(key=[ENTER YOUR API HERE])

efap = pd.read_csv('/home/nick/committees/SOC_HealthyNYC_2022/food_access/data/EFAP_pdf_4_7_22.csv', dtype=str) 
efap.head()

conditions = [
    (efap["DISTBORO"]=='BK'),
    (efap["DISTBORO"]=='BX'),
    (efap["DISTBORO"]=='NY'),
    (efap["DISTBORO"]=='QN'),
    (efap["DISTBORO"]=='SI'),
]
choices = ["Brooklyn", "Bronx", "Manhattan", "Queens", "Staten Island"]
efap["boro"] = np.select(conditions, choices)

efap['Address'] = efap['PROGRAM'] + ", " + efap['DISTADD'] + ", " + efap['boro'] + ", " + efap["DISTZIP"] + ", NY"

efap.head()

addresses = efap['Address'].unique()
len(addresses)

lat=[]
lng=[]
# loop through each address to search
for i, address in enumerate(addresses):
    #use the google maps interface to search for each address
    geocode_result = gmaps.geocode(address)
    #extract latitude & longitude and add to list of results
    if geocode_result:
        lat.append(geocode_result[0]['geometry']['location']['lat'])
        lng.append(geocode_result[0]['geometry']['location']['lng'])
    else: 
        lat.append('NA')
        lng.append('NA')

# create a table of results (address, lat, lon)
df_locations = pd.DataFrame(addresses).rename(columns = {0:'Address'})

df_locations['lat'] = lat
df_locations['lon'] = lng

df_locations

df_locations.to_csv('/home/nick/committees/SOC_HealthyNYC_2022/food_access/data/address_lat_lon.csv', index=False)

efap_merged = efap.merge(df_locations, how='left', on='Address')
# remove full address columns
efap_merged = efap_merged.loc[:, efap_merged.columns != "Address"]

efap_merged

efap_merged.to_csv('/home/nick/committees/SOC_HealthyNYC_2022/food_access/data/efap_geo.csv', index=False)
